[Telink] Fix CI build

+++ scripts/tools/telink/process_binaries.py
@@ -51,7 +51,8 @@ def merge_binaries(input_file1, input_file2, output_file, offset):
 build_conf = BuildConfiguration(os.path.join(os.getcwd(), os.pardir))
 
 # Clean up merged.bin from previuse build
-os.remove('merged.bin')
+if os.path.exists('merged.bin'):
+    os.remove('merged.bin')
 
 # Telink W91 dual-core SoC binary operations
 if build_conf.getboolean('CONFIG_SOC_SERIES_RISCV_TELINK_W91'):
